Add filtereing and view toggling
Change-type: minor
