+++ package.json
@@ -23,6 +23,7 @@
     "@nuxtjs/dotenv": "^1.4.0",
     "@nuxtjs/pwa": "^3.2.2",
     "@nuxtjs/sitemap": "^2.4.0",
+    "@nuxtjs/toast": "^3.3.1",
     "@sindresorhus/slugify": "^1.1.0",
     "@tiptap/core": "^2.0.0-alpha.7",
     "@tiptap/extension-collaboration": "^2.0.0-alpha.4",
